+++ src/FreeElephants/RestDaemon/Endpoint/Handler/EndpointMethodHandlerInterface.php
@@ -8,19 +8,13 @@ use FreeElephants\RestDaemon\Middleware\MiddlewareInterface;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
 use Psr\Http\Message\UriInterface;
+use Psr\Http\Server\RequestHandlerInterface;
 
 /**
  * @author samizdam <samizdam@inbox.ru>
  */
-interface EndpointMethodHandlerInterface extends MiddlewareInterface
+interface EndpointMethodHandlerInterface extends MiddlewareInterface, RequestHandlerInterface
 {
-
-    /**
-     * @deprecated will be make protected or private.
-     * @param ServerRequestInterface $request
-     * @return ResponseInterface
-     */
-    public function handle(ServerRequestInterface $request): ResponseInterface;
 
     public function setMiddlewareCollection(EndpointMiddlewareCollectionInterface $endpointMiddlewareCollection);